+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [keep a changelog](http://keepachangelog.com) and this pr
 - Adjust console path for delete all data operation.
 - Check HTTP key before Authorization header in RPC function invocations.
 - Surface error message from Apple authentication JWT verification failures.
+- Improve user online status lookups.
 
 ### Fixed
 - Fix data returned by StreamUserList in JS runtime.